# static/swagger.yaml
openapi: 3.0.0
info:
  title: Flask API
  description: API con operaciones CRUD documentada con Swagger
  version: 1.0.0
servers:
  - url: http://192.168.0.26:5000
    description: Servidor local

paths:
  /items:
    get:
      summary: Obtener todos los items
      responses:
        '200':
          description: Lista de todos los items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
    post:
      summary: Crear un nuevo item
      requestBody:
        description: Datos del nuevo item
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Nuevo item"
      responses:
        '201':
          description: Item creado xd

  /items/{id}:
    get:
      summary: Obtener un item por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        '404':
          description: Item no encontrado
    put:
      summary: Actualizar un item por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        description: Datos para actualizar el item
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Item actualizado"
      responses:
        '200':
          description: Item actualizado
        '404':
          description: Item no encontrado
    delete:
      summary: Eliminar un item por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item eliminado
        '404':
          description: Item no encontrado